plan2-as-plan - Plan2 => Plan
