+++ src/jsMain/kotlin/App.kt
@@ -12,7 +12,6 @@ enum class Label(val text: String) {
     Register("Register"),
     Organize("Organize"),
     Plan("Plan"),
-    Plan2("Plan2"),
     Prioritize("Prioritize")
 }
 
@@ -65,7 +64,6 @@ class App : RComponent<RProps, AppState>() {
                         mTab(Label.Register.text, Label.Register.text)
                         mTab(Label.Organize.text, Label.Organize.text)
                         mTab(Label.Plan.text, Label.Plan.text) //This can be personal or community
-                        mTab(Label.Plan2.text, Label.Plan2.text) //This can be personal or community
                         mTab(Label.Prioritize.text, Label.Prioritize.text) //This can be personal or community
                     }
                 }
@@ -76,10 +74,7 @@ class App : RComponent<RProps, AppState>() {
                     Label.Organize.text -> {
                         organize()
                     }
-                    Label.Plan.text -> {
-                        child(Plan) {}
-                    }
-                    Label.Plan2.text -> plan2 {
+                    Label.Plan.text -> plan2 {
                         chores = state.chores
                         deleteChore = { id ->
                             val scope = MainScope()
