+++ src/jsMain/kotlin/Plan2.kt
@@ -8,6 +8,11 @@ import react.*
 import styled.StyleSheet
 import styled.css
 import styled.styledDiv
+import kotlinx.html.InputType
+import org.w3c.dom.events.Event
+import org.w3c.dom.HTMLInputElement
+import react.dom.*
+import kotlinx.html.js.*
 
 external interface Plan2Props: RProps {
     var chores: List<Chore>
@@ -77,3 +82,32 @@ fun RBuilder.plan2(handler: Plan2Props.() -> Unit): ReactElement {
         this.attrs(handler)
     }
 }
+
+external interface InputProps : RProps {
+    var onSubmit: (String) -> Unit
+}
+
+val InputComponent = functionalComponent<InputProps> { props ->
+    val (text, setText) = useState("")
+
+    val submitHandler: (Event) -> Unit = {
+        it.preventDefault()
+        setText("")
+        props.onSubmit(text)
+    }
+
+    val changeHandler: (Event) -> Unit = {
+        val value = (it.target as HTMLInputElement).value
+        setText(value)
+    }
+
+    form {
+        attrs.onSubmitFunction = submitHandler
+        input(InputType.text) {
+            attrs.onChangeFunction = changeHandler
+            attrs.value = text
+        }
+    }
+}
+
+fun RBuilder.inputComponent(handler: InputProps.() -> Unit) = child(InputComponent) { attrs { handler() } }
